fix(webhook): extract resolveChannelTarget to node-cron-free leaf
The gateway bundle imported resolveChannelTarget transitively from
agent-scheduler/index.ts, which require()s node-cron. The gateway
build (bun build, no --external node-cron) then failed to resolve it,
breaking build-base/build-dist/e2e/race-long. Move resolveChannelTarget,
resolveEntryThreadId, and AgentChannelTarget into a pure leaf module
(channel-target.ts); index.ts re-exports for compatibility; the webhook
ingest path imports from the leaf so node-cron stays out of the gateway
bundle.
